+++ alicloud/service_alicloud_common.go
@@ -1,0 +1,59 @@
+package alicloud
+
+import (
+	"bytes"
+	"encoding/json"
+	"log"
+
+	"gopkg.in/yaml.v2"
+)
+
+func CompareJsonTemplateAreEquivalent(tem1, tem2 string) (bool, error) {
+	var obj1 interface{}
+	err := json.Unmarshal([]byte(tem1), &obj1)
+	if err != nil {
+		return false, err
+	}
+
+	canonicalJson1, _ := json.Marshal(obj1)
+
+	var obj2 interface{}
+	err = json.Unmarshal([]byte(tem2), &obj2)
+	if err != nil {
+		return false, err
+	}
+
+	canonicalJson2, _ := json.Marshal(obj2)
+
+	equal := bytes.Compare(canonicalJson1, canonicalJson2) == 0
+	if !equal {
+		log.Printf("[DEBUG] Canonical template are not equal.\nFirst: %s\nSecond: %s\n",
+			canonicalJson1, canonicalJson2)
+	}
+	return equal, nil
+}
+
+func CompareYmalTemplateAreEquivalent(tem1, tem2 string) (bool, error) {
+	var obj1 interface{}
+	err := yaml.Unmarshal([]byte(tem1), &obj1)
+	if err != nil {
+		return false, err
+	}
+
+	canonicalYaml1, _ := yaml.Marshal(obj1)
+
+	var obj2 interface{}
+	err = yaml.Unmarshal([]byte(tem2), &obj2)
+	if err != nil {
+		return false, err
+	}
+
+	canonicalYaml2, _ := yaml.Marshal(obj2)
+
+	equal := bytes.Compare(canonicalYaml1, canonicalYaml2) == 0
+	if !equal {
+		log.Printf("[DEBUG] Canonical template are not equal.\nFirst: %s\nSecond: %s\n",
+			canonicalYaml1, canonicalYaml2)
+	}
+	return equal, nil
+}